+++ src/lazyLoad/index.ts
@@ -283,6 +283,9 @@ const lazyLoad = async (
                         fs.writeFileSync(path.join(output, new URL(url).host.replace(":", "_") + "/BUILD_ID"), buildId);
                     }
                 }
+
+                // extract the source maps
+                await extractSourceMaps(output, sourcemapDir);
             } else if (tech.name === "vue") {
                 console.log(chalk.green("[✓] Vue.js detected"));
                 console.log(chalk.yellow(`Evidence: ${tech.evidence}`));